Make blurlock to turn off the screen after 15s
As of version 2.8 i3lock does not support DPMS settings. Instead, they added a handy script to the man page.
This is basically an integration of the script from the man page and blurlock. For me, this is constantly saving a lot of battery on the laptop.

Tested only with one display. It's said that script from man page should be working correctly for multiple displays.
See https://www.mankier.com/1/i3lock#Dpms